Restructured Database.php to handle different connection styles.
A few things happening. Comment insertion is broken right now because
apostrophes aren't handled properly because the original is stupid.

If I'm going to fix a query I'm going to move towards the class-based
structure I have in mind, which means a bit of background work is
necessary. I can't just force it into using a prepared statement with
mysqli (which is what I tried doing in previous commits) because there
are different connection mechanisms vs old style mysql_connect() sort of
stuff.

So: Database.php now handles different connection types. The vast
majority of the site will never encounter it but post.php will be
updated to use the new stuff.

Anyway it's not super great, a lot of dupey dupey, but at least now
instances of the site won't need to manage both Database.php and
include.php.